Add restore bar mode
Restore the mode of the bar to the value it has been when focus mode was
started.
Add call to get the bar mode.
Add function to restore mode. It will only restore to the previous mode
if no change in the mode was detected. Assuming the user change the mode
while the focus timer was running, don't change it to something else.